Massive aesthetics update outlined below;
1) Finally fixed the corners of the cards- they are now rounded as
originally intended. I had to round both the card AND the container in
the card.
2) As a test, instead of the tournament number, on the results page you
can see the tournament score.
3) Managed to animate the switch between the title and search text bar
after about a million attempts. The key was using AnimateCrossFade
widget. Simple and elegant solution.
4) I successfully centered the text in the drawer. The issue was
the icon always shifted the text over to the left even if wrapped in Center. I learned that every icon found
in Flutter is 24*24. Thus, I added 24px Padding, perfectly centering the
text.
5) Lastly, the drawer size seemed too big when loaded on my device. To
make sure this never happens, the drawer size should be modified per the
device. I decided to set the drawer width to 58% of the entire screen
width. The user can easily click out, and the drawer does not seem
overly huge.